same

+++ rlkit/torch/prg/prg.py
@@ -118,7 +118,7 @@ class PRGTrainer(TorchTrainer):
                 self.target_entropy = -np.prod(self.env.action_space.shape).item()  # heuristic value from Tuomas
             if self.use_entropy_loss:
                 # self.log_alpha_n = [ptu.zeros(1, requires_grad=True) for i in range(len(self.policy_n))]
-                self.log_alpha_n = [ptu.tensor([np.log(self.init_alpha)], requires_grad=True) for i in range(len(self.policy_n))]
+                self.log_alpha_n = [ptu.tensor([np.log(self.init_alpha)], requires_grad=True, dtype=torch.float32) for i in range(len(self.policy_n))]
                 self.alpha_optimizer_n = [
                     optimizer_class(
                         [self.log_alpha_n[i]],
@@ -127,7 +127,7 @@ class PRGTrainer(TorchTrainer):
 
             if self.use_cactor_entropy_loss:
                 # self.log_calpha_n = [ptu.zeros(1, requires_grad=True) for i in range(len(self.policy_n))]
-                self.log_calpha_n = [ptu.tensor([np.log(self.init_alpha)], requires_grad=True) for i in range(len(self.policy_n))]
+                self.log_calpha_n = [ptu.tensor([np.log(self.init_alpha)], requires_grad=True, dtype=torch.float32) for i in range(len(self.policy_n))]
                 self.calpha_optimizer_n = [
                     optimizer_class(
                         [self.log_calpha_n[i]],
